+++ packages/polar-nuxt/src/runtime/index.ts
@@ -1,7 +1,7 @@
 import {
   handleWebhookPayload,
   type WebhooksConfig,
-} from "@polar-sh/adapter-core";
+} from "@polar-sh/adapter-utils";
 import { Polar } from "@polar-sh/sdk";
 import {
   WebhookVerificationError,