+++ packages/polar-nextjs/CHANGELOG.md
@@ -15,14 +15,14 @@
 
 - 98a1bf9: add core dependency
 - Updated dependencies [98a1bf9]
-  - @polar-sh/adapter-core@0.1.3
+  - @polar-sh/adapter-utils@0.1.3
 
 ## 0.3.3
 
 ### Patch Changes
 
 - Updated dependencies [4abbedc]
-  - @polar-sh/adapter-core@0.1.2
+  - @polar-sh/adapter-utils@0.1.2
 
 ## 0.3.2
 
@@ -30,7 +30,7 @@
 
 - 46ad781: Await webhook handlers
 - Updated dependencies [46ad781]
-  - @polar-sh/adapter-core@0.1.1
+  - @polar-sh/adapter-utils@0.1.1
 
 ## 0.3.1